suit: Fix incorrect DFU cache sink slot size after seek
If the seek function of the DFU cache was used with
an offset smaller than the current offset, the filled
size of the slot would be interpreted incorrectly (it
could be smaller than the real size, it would show
the end of the written data after the last seek).

This commit fixes this issue so that the total size
of the slot is used.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>