mgmt: suitfu: unblock fetching from suit-processor
This commit unblocks fetching using SMP by the suit-processor
run on the application core. This is done by moving running
the processor out of the suitfu_mgmt_candidate_envelope_stored
function.

To fix other issues, instead of offloading the preprocessing
to the main system work queue, a separate worker thread with
a large enough stack size was added.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>