suit: source_selector based cache implementation
The previous implementation assumed that the NVM storage was directly
readable by the CPU. This is not true when a partition is located in
external memory.

The cache implementation now uses the source_selector module to pull the
data from the correct device when performing cache look-up and
modifications.

Added a set of helpers for iterating and pulling information out of the
cache partitions.

Added an adhoc sink implementation, which fulfills the sink contract and
allows for streaming data from an arbitrary source into a provided
buffer in memory.

Signed-off-by: Rafał Kuźnia <rafal.kuznia@nordicsemi.no>