suit: preinitialize suit_dfu_cache_rw module
When adding changes allowing to use SUIT DFU
cache with external flash translating addresses
of DFU cache partition is suit_dfu_cache_rw_initialize
was added. However, this function is not called in
the SUIT application orchestrator until the envelope
is stored.
The translated addresses of this partitions might
be needed before the envelope is stored - the user
might need to erase all the cache partitions before
he starts storing the envelope. In this case the
translated addresses will be needed to perform the erase.

This commit adds a prenitialize function to the
dfu_cache_rw module which translates the addresses
at startup. This function is called from a SYS_INIT.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>