suit: Refactor fetch source manager
Refactored the fetch source manager so that the application
does not have to link to internal SUIT interface libraries
nor use internal SUIT types such as the "sink" type.

The fetch source manager now provides interfaces
exposed to the user, which allow for registering fetch
sources and passing data fetched by the fetch sources
to the SUIT subsystem - note this is now done by dedicated
APIs instead of sinks.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>