[nrf noup] app: net: mdns improvement
In case when dname is not null termianted string (e.g. when buffer
is dirty where it has been copied) then `strncmp` is better to use.
Additionally the data length are compared earlier so only contetnt
should be checked.

The `iterate_mdns_records()` can be interrupted by callback (e.g.
after finding the searched record). This improve mDNS performance
because is no longer required to wait until last record is reached.

During processing question the query string is changed
to lowercase in order to provide it in unified form for the
function which search and compare available mDNS records.

Signed-off-by: Marcin Gasiorek <marcin.gasiorek@nordicsemi.no>