suit: change Kconfig dependencies for external memory sink/source
In the external memory service the roles between secure domain and the
application domain are reversed. The application domain provides a
service and the secure domain is the client. The kconfig dependencies
indicated the reverse, the SSF_SERVER_EXTMEM_SERVICE_ENABLED is enabled
on the application core and SSF_SERVER_EXTMEM_SERVICE_ENABLED is enabled
on the secure domain only. Changed the dependencies of the EXTMEM
sink/source enabling options to depend on
SSF_SERVER_EXTMEM_SERVICE_ENABLED.

Signed-off-by: Rafał Kuźnia <rafal.kuznia@nordicsemi.no>