suit: dfu_cache_rw: Verify size when iterating through slots
Added protection against a situation where bstr data in
a cache slot would exceed the remaining cache partition size.
Note this includes a situation where a cache slot was not
commited (an the size is 0xFFFFFFFF).

A partition containing such a slot should be treated as
invalid and recovered at cache initialization.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>