samples: crypto: persistant_key_usage: Add nrf54h20
Add nrf54h20 hack to support persistant keys.
It configures the secure storage used by the secure domain for its and
secure storage of the local domains to be in the uicr partition.
For this it just writes it straight into the UICR, so the UICR format is
also included.

Signed-off-by: Karsten Koenig <karsten.koenig@nordicsemi.no>